Many small adjustments, outlined below.
+ Moved the categories request into apps.js, and (temporarily) into the
problem controller. See the issue on gitlab.
+ Added a function into discussions.js for clearing tags
+ deleted a bit of debugging output that was left over from past
development
+ Fixed the notifications model to look at the 'read' column rather than
the 'seen' column. Had been throwing an exception prior.
+ Fixed threads to be loaded in a single request.
+ Fixed issue with duplicate shorthands being created. In the case of a
shorthand already used, a hash is created from the shorthand, the user
id, and the current datetime. Previously, it used just a hash of the
shorthand, which would be the same for any duplicates.
+ fixed the flag button to not reload the page on flagging
(preventdefault)
+ fixed the accumulation of content between loads of the problem page.
+ corrected the spelling of "contributors" ul on the problem page

Issues to be fixed:
On logout, the page redirects to a 404
To save ourself a request, have the problem load request the entire page
with just the problem shorthand, rather than requesting an id first and
then the problem info.